ToxicSense/requirements.txt
astroid==2.0.4
beautifulsoup4==4.6.3
bs4==0.0.1
certifi==2018.10.15
chardet==3.0.4
coala-utils==0.5.1
Django==2.1.2
idna==2.7
isort==4.3.4
lazy-object-proxy==1.3.1
lxml==4.2.5
mccabe==0.6.1
pylint==2.1.1
python-twitter
pytz==2018.6
requests==2.20.0
six==1.11.0
twitterscraper==0.9.0
urllib3==1.24
wrapt==1.10.11
Unidecode==1.0.22
Keras==2.2.4
Keras-Applications==1.0.6
Keras-Preprocessing==1.0.5
tensorboard==1.11.0
tensorflow==1.11.0
numpy==1.15.3
